Separate functions to utility module
