+++ src/rng/mod.rs
@@ -1,3 +1,4 @@
 pub mod bitmixer;
 pub mod pcg;
+pub mod util;
 pub mod xorshift;
